Adds QWERTY keyBoard and correct answer display
Reworks HTML layout to make room for keyBoard.

Adds text field above "Guess a word!" button to display the correct
answer if all guesses are exhausted and the correct word is not guessed.

Adds a function to construct a QWERTY, button based keyBoard below
the gameBoard.

Adds css formatting to keyBoard as well as small styling changes to the
rest of the page.

Adds function to color the keyBoard keys according to the entered guess.
Green for a match, orange for a close match, dark gray for a non match.

Updates README file to accurately reflect the current state of the
project.